Download _ActiveMQ_ from container
Currently the _ActiveMQ_ tarball is downloaded by the build script,
before building the container. The build script caches it in the
`.downloads` directory in order to avoid downloading it repeatedly. But
this directory is not used in the continuous integration environments.
For example, when building in Jenkins the tarball is downloaded
everytime. To avoid that this patch moves the download to the build of
the container. This way it will be cached as part of the container build
caching process. The net result is a couple of minutes less in most
builds, and a shorter build script.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>