Enable ActiveMQ console
Currently the _ActiveMQ_ container has the management console. This
patch enables it. The console will be available via the route for host
name `messaging-console`. For example, if the domain selected is
`example.com` then the console will be available in the following URL:

  https://messaging-console.example.com/console/

The default user name and password are `admin` and `redhat123`.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>