Add iSCSI target to the sandbox
Currently it isn't possible to use persistent volumes in the sandbox
virtual machine, because it doesn't have access to any external storage.
In particular it isn't possible to create a persistent volume to store
the data of the message broker. To allow that this patch configures an
iSCSI target in the virtual machine.

The target has ten LUNs, each of one GiB, backed by sparse files in the
`/storage` directory.

For each LUN there is a persitent volume.

The number and size of these LUNs and volumes can be changed with the
`sandbox_lun_count` and `sandbox_lun_size` in the `inventory.j2` file.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>