Add ActiveMQ Artemis launcher
Currently the configuration and instance files of the messaging broker
are pre-created during the creation of the image. This means that the
configuration parameters can't be changed without rebuilding the image.
This is also true for the passwords, the certificates and the private
keys used for the TLS endpoint. To avoid that this patch changes the
image so that it contains a new `artemis-launcher` tool that creates the
instance with a default configuration when the container is started, if
it doesn't already exist. If the container is started again with the
same data volume mounted, then the `artemis-launcher` tool will just
start the instance.

The `artemis-launcher` tool can also replace the configuration files
before starting the instance. This is intended to be able to have the
configuration files in a Kubernetes configuration map instead of having
them hard-coded in the image. The patch modifies the `template.yml` file
so that it uses this mechanism.

The `artemis-launcher` tool also has options that allow it to convert
TLS certificates and private keys in PEM format to the PKCS #12 format
supported by the messaging broker. This is combined with the _OpenShift_
capability to dynamically generate certificates and private keys.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>