Handle correctly zero customers
Currently when the result of retrieving a list of customers has zero
items the service tries anyhow to find the details, using the `in` SQL
mechanism:

```sql
select ... where id in ()
```

That causes an error which is then displayed in the document returned to
the user. This patch changes the service so that it doesn't try to run
this query when there are zero items.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>