Remove `build.py`
This patch removes the `build.py` script and replaces it with direct
invocation of the build tools from the `Makefile`.

The more relevant effect of this change is that now the project needs to
be checked out in a directory that is in the `GOPATH`. For example, if
your `GOPATH` is `/files/go` then you will have to do the following:

```
$ cd /files/go/src
$ mkdir -p github.com/container-mgmt
$ cd github.com/container-mgmt
$ git clone https://github.com/container-mgmt/dedicated-portal.git
$ cd dedicated-portal
$ make
```

The generated binaries and packages will be now in the `bin` and `pkg`
sub-directories of your `GOPATH`, and will not be removed with `make
clean`.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>