Fix Go environment in Jenkinsfile
After removing the `build.py` script it isn't possible to build the Go
parts of the project outside of the directory pointed by the `GOPATH`
environment variable, and the Jenkins build is failing. To solve that
this patch changes the `Jenkinsfile` so that `GOPATH` points to the
workspace, and so that the source code is checked out in the
subdirectory that the Go tools expect.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>