Always change to the `dedicated-portal` project
Currently the `template.sh` script and the `Jenkinsfile` try to create
the `dedicated-portal` project if it doesn't exist. But if the project
already exists, and it isn't the current one, they don't make it the
current one, and as a result the objects may be created in other
namespace. This patch changes both files so that if the creation of the
project fails (because it already exists, most probably) they try to
make it the current one explicitly.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>