Move API types to `api` package
Currently the types used in the API are defined in the same packages
where the servers are implemented. This makes it easy to hink that those
types can be tweaked to make life simpler for the implementations of the
servers. To avoid that this patch moves them to a new `api` package, so
it is clearer that these types are part of the contract of the services,
and that they need to be kept stable and backwards compatible.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>