Fix the pods CIDR block
Currently when we create a `clustersdeployment` object to request the
creation of a cluster to the cluster operator we se the following CIDR
blocks:

- Services: `172.30.0.0/16`
- Pods: `172.30.0.0/14`

Teh first is correct, but the second is an incorrect CIDR, for a 14 bits
mask it should have been `172.28.0.0/14`. This problem is detected by
the the master controllers pod of the created cluster, which reports
this in its log:

```
E0803 18:51:58.596416 1 common.go:55]
Configured clusterNetworks value "172.30.0.0/14" is invalid;
treating it as "172.28.0.0/14"
```

The cluster tries then to use that CIDR block, but it conflicts with the
CIDR block used by the `us-east-1` zone. That is also detected and
reported th the master controller pod of the created cluster:

```
F0803 18:51:58.597869 1 controller_manager.go:194]
Error starting "openshift.io/sdn"(failed to start SDN plugin controller:
cluster IP: 172.28.0.0 conflicts with host network: 172.31.48.0/20)
```

This is a fatal error, so the master controllers pod doesn't start, and
then the installation of the rest of the cluster also fails.

To address this issue this patch changes the CIDR block to use one that
doesn't conflict, the same that was used in a similar change in the
examples of the cluster operator:

> Fix cluster deployment template subnets
> https://github.com/openshift/cluster-operator/commit/6cf7ec6b4fd1b3038856742641a02ce2b54e84e5

Signed-off-by: Juan Hernandez <jhernand@redhat.com>