Pull cluster operator images if not present
Currently the specification of the cluster version used with the cluster
operator specifies that images should never be pulled. This is OK for
development environments, where the images are pulled manually, but it
is not OK for production environments. This patch changes the
specifiation of the cluster version so that images are pulled if not
present.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>