Add API endpoint to get cluster status
Add `.../clusters/{id}/status` endpoint to get the cluster status.
Added a label "ID" with the cluster id in the `clusterdeployment`
This ID is used to find the cluster when status is requested.

Another important change: first create the cluster in the DB, and then try provisioning it if required.
Also add ClusterRole and ClusterRoleBinding creation in template.yml to allow the uhc sa to access clusteroperator CRDs